requirements.txt
pdfminer3
pyresparser
streamlit
pandas
pafy
plotly
streamlit-tags
Pillow
youtube-dl
nltk
pdfminer3
